Address PR review: restore configs to main.py, improve typing, add TimeSeriesLogs protocol
- Move StepMeanEntry, InferenceEvaluatorAggregatorConfig, InferenceAggregatorConfig
  back to main.py at original positions; delete config.py
- Type _aggregators as dict[str, SubAggregator] instead of dict[str, Any]
- Add TimeSeriesLogs protocol for step_slice-capable get_logs; replace
  _streaming_names list with _time_series_aggregators dict
- Remove @runtime_checkable from SubAggregator
- Pass time in record_initial_condition instead of None
- Restore deleted docstrings
- Add InferenceBatchData.new_test_data() classmethod for test convenience
